miner/requirements.txt
gitpython==3.1.41
tree-sitter==0.21.3
redis==5.0.1
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
